Trait bonsaidb_files::FileConfig

pub trait FileConfig: Sized + Send + Sync + Unpin + 'static {
    const BLOCK_SIZE: usize;

    fn files_name() -> CollectionName;

    fn blocks_name() -> CollectionName;

    fn register_collections(schema: &mut Schematic) -> Result<(), Error> { ... }

    fn build<NameOrPath: AsRef<str>>(
        name_or_path: NameOrPath
    ) -> FileBuilder<'static, Self> { ... }

    fn get<Database: Connection + Clone>(
        id: u32, 
        database: Database
    ) -> Result<Option<File<Blocking<Database>, Self>>, Error> { ... }

    fn load<Database: Connection + Clone>(
        path: &str, 
        database: Database
    ) -> Result<Option<File<Blocking<Database>, Self>>, Error> { ... }

    fn list<Database: Connection + Clone>(
        path: &str, 
        database: &Database
    ) -> Result<Vec<File<Blocking<Database>, Self>>, Error> { ... }

    fn list_recursive<Database: Connection + Clone>(
        path: &str, 
        database: &Database
    ) -> Result<Vec<File<Blocking<Database>, Self>>, Error> { ... }

    fn get_async<'async_trait, Database>(
        id: u32, 
        database: Database
    ) -> Pin<Box<dyn Future<Output = Result<Option<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>>
    where
        Database: 'async_trait + AsyncConnection + Clone,
        Self: 'async_trait,
    { ... }

    fn load_async<'life0, 'async_trait, Database>(
        path: &'life0 str, 
        database: Database
    ) -> Pin<Box<dyn Future<Output = Result<Option<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>>
    where
        Database: 'async_trait + AsyncConnection + Clone,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn list_async<'life0, 'life1, 'async_trait, Database>(
        path: &'life0 str, 
        database: &'life1 Database
    ) -> Pin<Box<dyn Future<Output = Result<Vec<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>>
    where
        Database: 'async_trait + AsyncConnection + Clone,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn list_recursive_async<'life0, 'life1, 'async_trait, Database>(
        path: &'life0 str, 
        database: &'life1 Database
    ) -> Pin<Box<dyn Future<Output = Result<Vec<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>>
    where
        Database: 'async_trait + AsyncConnection + Clone,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

A configuration for a set of stored files.

Associated Constants

const BLOCK_SIZE: usize

The maximum size for each write to an underlying file. The file will be stored by breaking the data written into chunks no larger than BLOCK_SIZE.

Required methods

fn files_name() -> CollectionName

Returns the unique collection name to use to store Files.

fn blocks_name() -> CollectionName

Returns the unique collection name to use to store file blocks.

Provided methods

fn register_collections(schema: &mut Schematic) -> Result<(), Error>

Registers the collections for this configuration into schema.

fn build<NameOrPath: AsRef<str>>(
name_or_path: NameOrPath
) -> FileBuilder<'static, Self>

Builds a new file. If name_or_path starts with a /, the argument is treated as a full path to the file being built. Otherwise, the argument is treated as the file’s name.

fn get<Database: Connection + Clone>(
id: u32,
database: Database
) -> Result<Option<File<Blocking<Database>, Self>>, Error>

Returns the file with the unique id given, if found. This function only loads metadata about the file, it does not load the contents of the file.

fn load<Database: Connection + Clone>(
path: &str,
database: Database
) -> Result<Option<File<Blocking<Database>, Self>>, Error>

Returns the file located at path, if found. This function only loads metadata about the file, it does not load the contents of the file.

fn list<Database: Connection + Clone>(
path: &str,
database: &Database
) -> Result<Vec<File<Blocking<Database>, Self>>, Error>

Returns all files that have a containing path of exactly path. It will only return files that have been created, and will not return “virtual” directories that are part of a file’s path but have never been created.

This function only loads metadata about the files, it does not load the contents of the files.

fn list_recursive<Database: Connection + Clone>(
path: &str,
database: &Database
) -> Result<Vec<File<Blocking<Database>, Self>>, Error>

Returns all files that have a path starting with path.

This function only loads metadata about the files, it does not load the contents of the files.

fn get_async<'async_trait, Database>(
id: u32,
database: Database
) -> Pin<Box<dyn Future<Output = Result<Option<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>> where
Database: 'async_trait + AsyncConnection + Clone,
Self: 'async_trait,

Returns the file with the unique id given, if found. This function only loads metadata about the file, it does not load the contents of the file.

fn load_async<'life0, 'async_trait, Database>(
path: &'life0 str,
database: Database
) -> Pin<Box<dyn Future<Output = Result<Option<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>> where
Database: 'async_trait + AsyncConnection + Clone,
'life0: 'async_trait,
Self: 'async_trait,

Returns the file located at path, if found. This function only loads metadata about the file, it does not load the contents of the file.

fn list_async<'life0, 'life1, 'async_trait, Database>(
path: &'life0 str,
database: &'life1 Database
) -> Pin<Box<dyn Future<Output = Result<Vec<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>> where
Database: 'async_trait + AsyncConnection + Clone,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Returns all files that have a containing path of exactly path. It will only return files that have been created, and will not return “virtual” directories that are part of a file’s path but have never been created.

This function only loads metadata about the files, it does not load the contents of the files.

fn list_recursive_async<'life0, 'life1, 'async_trait, Database>(
path: &'life0 str,
database: &'life1 Database
) -> Pin<Box<dyn Future<Output = Result<Vec<File<Async<Database>, Self>>, Error>> + Send + 'async_trait>> where
Database: 'async_trait + AsyncConnection + Clone,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Returns all files that have a path starting with path.

This function only loads metadata about the files, it does not load the contents of the files.